+++ src/main/resources/com/juego/application/equipo4juegotrivia/resultados.fxml
@@ -1,18 +1,26 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.VBox?>
-<VBox alignment="CENTER" spacing="20" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
-    <Label text="¡Juego Terminado!" style="-fx-font-size: 24px; -fx-font-weight: bold;"/>
-    <VBox alignment="CENTER" spacing="10">
-        <Label fx:id="puntuacionFinalLabel" style="-fx-font-size: 20px; -fx-text-fill: #2E7D32;"/>
-        <Label fx:id="respuestasCorrectasLabel" style="-fx-font-size: 16px;"/>
-        <Label fx:id="respuestasIncorrectasLabel" style="-fx-font-size: 16px;"/>
-        <Label fx:id="porcentajeAciertosLabel" style="-fx-font-size: 16px;"/>
-    </VBox>
+<?import javafx.scene.text.Font?>
 
-    <HBox fx:id="estrellasContainer" alignment="CENTER" spacing="10"/>
+<VBox alignment="CENTER" prefHeight="600.0" prefWidth="800.0" spacing="20.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.juego.application.equipo4juegotrivia.Jugador">
+    <children>
+        <Label text="Resultados Finales" style="-fx-font-weight: bold;">
+            <font>
+                <Font size="24.0" />
+            </font>
+        </Label>
 
-    <Button fx:id="volverMenuBtn" text="Volver al Menú Principal" style="-fx-font-size: 14px;"/>
+        <Label fx:id="puntuacionFinalLabel" text="Puntuación Final: 0" />
+        <Label fx:id="respuestasCorrectasLabel" text="Respuestas Correctas: 0/0" />
+        <Label fx:id="respuestasIncorrectasLabel" text="Respuestas Incorrectas: 0/0" /> <!-- ✅ Cambiado para mostrar /total -->
+        <Label fx:id="porcentajeAciertosLabel" text="Porcentaje de aciertos: 0%" />
+
+
+
+
+        <Button fx:id="volverMenuBtn" mnemonicParsing="false" text="Volver al Menú Principal" />
+    </children>
 </VBox>
